requirements.txt
absl-py==1.3.0
alembic==1.8.1
altair==4.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
bqplot==0.12.36
branca==0.6.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colour==0.1.5
commonmark==0.9.1
cycler==0.11.0
databricks-cli==0.17.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docker==6.0.1
entrypoints==0.4
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
flatdict==4.0.1
folium==0.13.0
fonttools==4.38.0
gdown==4.5.3
geojson==2.5.0
gitdb==4.0.9
GitPython==3.1.29
greenlet==2.0.0.post0
gunicorn==20.1.0
haversine==2.7.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.16.2
ipyleaflet==0.17.2
ipysheet==0.6.0
ipython==7.34.0
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.0
jupyter-server==1.23.1
jupyter_client==7.4.5
jupyter_core==4.11.2
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.3
kiwisolver==1.4.4
leafmap==0.12.1
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mlflow==1.30.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
oauthlib==3.2.2
ortools==9.4.1874
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.0
pyarrow==10.0.0
pycparser==2.21
PyCRS==1.0.2
pydeck==0.8.0
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pyshp==2.3.1
PySocks==1.7.1
pystac==1.5.0
pystac-client==0.5.1
pytest==7.1.3
python-box==6.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
querystring-parser==1.2.4
requests==2.28.1
rich==12.6.0
scikit-learn==1.0.2
scipy==1.7.3
scooby==0.7.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
sqlparse==0.4.3
streamlit==1.14.0
streamlit-folium==0.7.0
tabulate==0.9.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
traittypes==0.2.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
whitebox==2.2.0
whiteboxgui==2.2.0
widgetsnbextension==4.0.3
xyzservices==2022.9.0
zipp==3.9.0
